web7/requirements.txt
Flask
openai==0.27.0
pytesseract
PyPDF2
pandas
pillow
beautifulsoup4
